src/main/java/com/kolmakova/rest/EmployeeController.java
package com.kolmakova.rest;

import com.kolmakova.dto.Employee;
import com.kolmakova.service.EmployeeService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/employees")
public class EmployeeController {

    @Autowired
    public EmployeeService employeeService;

    @GetMapping
    public List<Employee> getAll() {
        return employeeService.getAll();
    }

    @GetMapping("{id}")
    public Employee getOne(@PathVariable Integer id) {
        return employeeService.getOne(id);
    }

    @PostMapping
    public Employee create(@RequestBody Employee employee) {
        return employeeService.save(employee);
    }

    @PutMapping
    public Employee update(@RequestBody Employee employee) {
        return employeeService.update(employee);
    }

    @DeleteMapping("{id}")
    public void delete(@PathVariable Integer id) {
        employeeService.delete(id);
    }
}
